maintenance

+++ code/__main__.cpp
@@ -51,14 +51,14 @@ int main()
 	joueur1.init(noir) ;
 	joueur2.init(blanc) ;
 
-  	fstream fich ;
-  	fich.open("resultats.dat", ios::out) ;
+  fstream fich ;
+  fich.open("resultats.dat", ios::out) ;
 
-  	int points_noir, points_blanc, gagnant = 0 ;					// gagnant: 0 = égalité ; 1 = blanc a gagné ; 2 = noir a gagné
+  int points_noir, points_blanc, gagnant = 0 ;					// gagnant: 0 = égalité ; 1 = blanc a gagné ; 2 = noir a gagné
 
-	for (int i=0; i<3; i++)
+	for (size_t i = 0 ; i < 3 ; i++)
 	{
-		points_noir = 0 ; points_blanc = 0 ; 
+		points_noir = 0 ; points_blanc = 0 ;
 		grille_de_jeu.init() ;
 
 		int coup_x = 0, coup_y = 0 ;
@@ -83,7 +83,7 @@ int main()
 		fich << points_blanc << ' ' << points_noir << ' ' << gagnant << endl ;
 	}
 
-  	fich.close();
+  fich.close() ;
 
 	return 0 ;
 }
